fix: resolve 3 MAJOR + 2 MINOR review findings across all phases
F-02 MAJOR (DRY): consolidate 3 duplicate Value→JSON conversion functions
into single Value::to_json() method on the AST enum. Delete field_value_to_json
(parser/dml.rs) and value_to_json (ddl_executor.rs).

F-03 MAJOR (CC): reduce validate_with_config from CC~12 to CC~5.
Extract requires_select_validation() helper. Merge duplicate compound
query validation loops into single pass.

F-11 MAJOR (Cohesion): extract DML mutation methods from ddl_executor.rs
(729 lines) into new dml_executor.rs. ddl_executor.rs now ~337 lines,
dml_executor.rs ~392 lines. Both under 500 NLOC.

F-01 MINOR: remove stale duplicate doc comment on execute_dml.
F-20 MINOR: add rustdoc on pub struct QueryValidator.

All 190 BDD + 1029 unit tests passing. Zero regressions.
